Fixed bug where using comms like i2c in an ISR could while forever
Some comm stuff relies on the clocking library to tell when it's timed out. But the clocking library only updates when it interrupts. So if the comm stuff is sitting in its own ISR it's likely the same priority interrupt, and thus won't ever let the clock update. Fixed by making clock high priority